ai_private_for_internal_docs/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
requests==2.32.3
qdrant-client==1.11.3
sentence-transformers==3.0.1
numpy==2.0.2

pypdf==4.3.1
python-docx==1.1.2
markdown==3.6
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0

# v3: OIDC/JWT
python-jose[cryptography]==3.3.0
